using contexts, router and middlewares
Using gorilla/context as a KV store for additional request information like
JRPCMethod that is being requested, this saves us from serializing the request
on every middleware hop.

gorilla/router is used to add middleware, and additionally can serve as a way
to seperate methods and middlewares - f.e. no authentication and logging
on /healthz

For middlewares we incluide augmentation (contexts), basic logging and telemetry
middlewares